refactor: 소셜로그인 뷰를 하나의 폴더 social_views로 묶음
