+++ user/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 from rest_framework_simplejwt.views import TokenRefreshView, TokenVerifyView
-from .google_login import GoogleLoginView, GoogleCallbackView, GoogleLoginToDjango
-from .kakao_login import KakaoLoginView, KakaoCallbackView, KakaoLoginToDjango
-from .github_login import GithubLoginView, GithubCallbackView, GithubLoginToDjango
+from user.social_views.google_login import GoogleLoginView, GoogleCallbackView, GoogleLoginToDjango
+from user.social_views.kakao_login import KakaoLoginView, KakaoCallbackView, KakaoLoginToDjango
+from user.social_views.github_login import GithubLoginView, GithubCallbackView, GithubLoginToDjango
 from .views import LogoutAPIView
